third_party/blink/web_tests/external/wpt/webrtc/RTCRtpReceiver-track-settings.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] RTCRtpReceiver.track.getSettings() contains resolution entries
  assert_equals: Receive track width setting is zero initially expected (number) 0 but got (undefined) undefined
Harness: the test ran to completion.
